Create Bootstrap.ini
MDT "ZTI" - settings for "Zero Touch Imaging". These settings make it zero touch imaging, you can pick the task to install and name the computer what you want or leave the default generated name. If you want to revert to a semi ZTI, just make a couple changes to the custom settings to assign the task, change YES to NO for the following: "SkipTask…